move processQ.await inside file loop
can't call processQ.await before it loads up the files, so need to move
it inside of the loop.  I have no idea why this ran on my machine
earlier, but it did, and then it failed to run on the ARB machine.

Also dropped fileQ because it is just a one-operation no need to use
queue to handle it, just send the callback to glob directly.